+++ flink-python/pyflink/datastream/connectors/tests/test_kafka.py
@@ -26,15 +26,19 @@ from pyflink.common.serialization import SimpleStringSchema, DeserializationSche
     JsonRowDeserializationSchema, CsvRowDeserializationSchema, AvroRowDeserializationSchema, \
     JsonRowSerializationSchema, CsvRowSerializationSchema, AvroRowSerializationSchema
 from pyflink.common.typeinfo import Types
-from pyflink.common.types import Row, to_java_data_structure
+from pyflink.common.types import Row
 from pyflink.common.watermark_strategy import WatermarkStrategy
 from pyflink.datastream.connectors.base import DeliveryGuarantee
 from pyflink.datastream.connectors.kafka import KafkaSource, KafkaTopicPartition, \
     KafkaOffsetsInitializer, KafkaOffsetResetStrategy, KafkaRecordSerializationSchema, KafkaSink, \
     FlinkKafkaProducer, FlinkKafkaConsumer
 from pyflink.java_gateway import get_gateway
-from pyflink.testing.test_case_utils import PyFlinkStreamingTestCase, PyFlinkTestCase, \
-    invoke_java_object_method
+from pyflink.testing.test_case_utils import (
+    PyFlinkStreamingTestCase,
+    PyFlinkTestCase,
+    invoke_java_object_method,
+    to_java_data_structure,
+)
 from pyflink.util.java_utils import to_jarray, is_instance_of, get_field_value
