+++ flink-python/pyflink/fn_execution/coder_impl_fast.pxd
@@ -194,10 +194,16 @@ cdef class AvroCoderImpl(FieldCoderImpl):
     cdef object _writer
 
 cdef class LocalDateCoderImpl(FieldCoderImpl):
-    pass
+    @staticmethod
+    cdef _encode_to_stream(value, OutputStream out_stream)
+    @staticmethod
+    cdef _decode_from_stream(InputStream in_stream)
 
 cdef class LocalTimeCoderImpl(FieldCoderImpl):
-    pass
+    @staticmethod
+    cdef _encode_to_stream(value, OutputStream out_stream)
+    @staticmethod
+    cdef _decode_from_stream(InputStream in_stream)
 
 cdef class LocalDateTimeCoderImpl(FieldCoderImpl):
     pass
